ConnectionFactory arrumado !!!

+++ dados-parlamentares/src/main/java/br/uel/bd1/dadosparlamentares/business/DeputadoBusiness.java
@@ -15,7 +15,7 @@ public class DeputadoBusiness extends GenericBusiness<Deputado> {
     protected CellProcessor[] getProcessors() {
         final CellProcessor[] processors = new CellProcessor[] {
                 new NotNull(new ParseLong()),
-                new NotNull(new ParseInt())
+                new NotNull(new ParseLong())
         };
 
         return processors;
